fix ci and add attach
